+++ src/main/dll/dll_0133_sfxplayer.c
@@ -56,7 +56,7 @@ void sfxplayerObj_init(u8* obj, u8* data)
     case SFXPLAYER_MODE_RANDOM_DELAY:
         {
             int v = randomGetRange(data[0x1e], data[0x1f]);
-            f32 fv = (f32)v;
+            f32 fv = v;
             fv = lbl_803E40BC * fv;
             *(f32*)sub = fv;
             break;
